+++ test/index.js
@@ -1,4 +1,2 @@
-
-
 require('./utils');
 require('./report');