+++ test/report.js
@@ -1,26 +1,26 @@
-import should from 'should';
+import path from 'path';
 
+import should from 'should';
 import _ from 'lodash';
 
 import {report} from '../src/report';
 
-import path from 'path';
-
 describe('report functions', () => {
-    it('successful report', (done) => {
+    it('successful report', done => {
+        report(
+            {directory: path.join(__dirname, 'assets', 'successful-report')},
+            result => {
+                const packageNames = _.map(result, 'name');
 
-        report({directory: path.join(__dirname, 'assets', 'successful-report')}, (result) => {
+                should(packageNames)
+                    .containEql('package-a')
+                    .and.containEql('package-b')
+                    .and.containEql('abbrev')
+                    .and.containEql('package-license')
+                    .and.not.containEql('package-dev');
 
-            const packageNames = _.map(result, 'name');
-
-            should(packageNames)
-                .containEql('package-a')
-                .and.containEql('package-b')
-                .and.containEql('abbrev')
-                .and.containEql('package-license')
-                .and.not.containEql('package-dev');
-
-            done();
-        });
+                done();
+            }
+        );
     });
 });