+++ test/utils.js
@@ -4,33 +4,36 @@ import {excludeRepositories, checkDependency} from '../src/util';
 
 const dependencies = [
     {
-        repository: 'https://gitlab.eccenca.com/elds-ui/ecc-mixins'
+        repository: 'https://gitlab.eccenca.com/elds-ui/ecc-mixins',
     },
     {
         name: 'ecc-component-dataintegrationwindow',
-        repository: 'ssh://git@gitlab.eccenca.com:8101/elds-ui/ecc-component-dataintegrationwindow'
+        repository:
+            'ssh://git@gitlab.eccenca.com:8101/elds-ui/ecc-component-dataintegrationwindow',
     },
     {
-        repository: 'https://github.com/elds/ecc-mixins'
+        repository: 'https://github.com/elds/ecc-mixins',
     },
     {
-        repository: 'https://github.com/component/emitter'
+        repository: 'https://github.com/component/emitter',
     },
     {
-        repository: 'ssh:fooo/asd/.git'
+        repository: 'ssh:fooo/asd/.git',
     },
     {
-        repository: undefined
+        repository: undefined,
     },
     {
-        repository: false
+        repository: false,
     },
 ];
 
 describe('utility functions', () => {
     describe('excludeRepositories', () => {
-
-        const testFn1 = excludeRepositories.bind(null, /(gitlab.eccenca.com)|(github.com\/elds\/)/);
+        const testFn1 = excludeRepositories.bind(
+            null,
+            /(gitlab.eccenca.com)|(github.com\/elds\/)/
+        );
         const testFn2 = excludeRepositories.bind(null, /git@gitlab/);
 
         it('should exclude eccenca dependencies (gitlab/gh)', () => {
@@ -55,14 +58,14 @@ describe('utility functions', () => {
         });
     });
 
-    describe('checkDependency', ()=> {
-
+    describe('checkDependency', () => {
         it('should throw error if a name or version is missing', () => {
-            should(function(){checkDependency(dependencies[0])}).throw(Error);
-            should(function(){checkDependency(dependencies[1])}).throw(Error);
-        })
-
+            should(() => {
+                checkDependency(dependencies[0]);
+            }).throw(Error);
+            should(() => {
+                checkDependency(dependencies[1]);
+            }).throw(Error);
+        });
     });
-
 });
-