+++ CHANGELOG.md
@@ -2,7 +2,16 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/) and [Keep A Changelog's Format](http://keepachangelog.com/).
 
-## [2.5.2] 2017-01-11 
+## [2.6.0] 2017-08-30
+
+### Added
+- Read files starting with `patent` (case insensitive) as notice files.
+- Read files starting with `copying` (case insensitive) as license files.
+
+### Changed
+- Update spdx license list to version 2.6
+
+## [2.5.2] 2017-01-11
 ### Fixed
 - `consolidate` allows downloadUrl and copies all fields if possible from input file
 - `yaml2json` now converts yaml correctly to json
@@ -38,7 +47,7 @@ DataManager:
 
 ## [2.4.0] 2016-12-12
 ### Added
-- possibility to provide a `additionalLicenses.yml` file which will be merged into the report 
+- possibility to provide a `additionalLicenses.yml` file which will be merged into the report
 
 ### Changed
 - Updated dependencies